% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fyleAttachmentGrabber.R
\name{fyleAttachmentGrabber}
\alias{fyleAttachmentGrabber}
\title{Retrieve Attachments from Email}
\usage{
fyleAttachmentGrabber(
  subject_keyword,
  max_recent_emails = 1000,
  attachment_folder = "saved_attachment_fyles"
)
}
\arguments{
\item{subject_keyword}{Search string or regex, though regex is not thoroughly tested. \emph{fyleAttachmentGrabber()} will search through your email inbox and pull all attachments from any email that has this string in its subject.}

\item{max_recent_emails}{How many emails to look at.  \emph{fyleAttachmentGrabber()} will search through \emph{max_recent_emails} or the number of emails in your inbox, whichever is less.}

\item{attachment_folder}{Optional name of folder to save attachments to.  This folder will be created in your current directory.  Specifying a directory with a full path may work but is untested.}
}
\description{
This function saves all attachments from emails containing specific text in the subject.  Works with COM email clients.
}
\keyword{extension}
\keyword{files,}
\keyword{filesearch,}
\keyword{find,}
