% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fyleGluer.R
\name{fyleGluer}
\alias{fyleGluer}
\title{Glue Together Email Attachments Split by fyleSplitter()}
\usage{
fyleGluer(attachment_folder = "saved_attachment_fyles")
}
\arguments{
\item{attachment_folder}{Optional name of folder to save attachments to.  This folder will be created in your current directory.  Specifying a directory with a full path may work but is untested.}
}
\description{
This function reconstructs emails that have been split by fyleSplitter() and deletes the individual splits.
}
\keyword{attachment}
\keyword{email,}
