% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fyleSender.R
\name{fyleSender}
\alias{fyleSender}
\title{Send Email with Attachments via Outlook}
\usage{
fyleSender(
  email_to,
  email_attachments = NULL,
  email_subject = "subject here",
  email_body = "body here",
  email_cc = "",
  email_bcc = ""
)
}
\arguments{
\item{email_to}{Email address to send the files to}

\item{email_attachments}{character vector of location of attachments to send. \emph{fyleIdentifier()$file} provides suitable input.}

\item{email_subject}{Optional email subject. Should have something unique about it so you can search through it later using fyleAttachmentGrabber()'s \emph{subject_keyword} argument.}

\item{email_body}{Optional email body}

\item{email_cc}{Optional email cc}

\item{email_bcc}{Optional email bcc}
}
\description{
Uses RDCOMClient package.
}
\keyword{attachments,}
\keyword{email,}
\keyword{outlook}
