% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fysan - orig.R, R/fysan.R
\name{fysan}
\alias{fysan}
\title{Batch Send Filetypes via Outlook Email}
\usage{
fysan(
  batch = 3,
  interval = 1,
  fyle_location = ".",
  fyle_extension = NULL,
  fyle_exclusions = NULL,
  list_recursive = TRUE,
  list_full_names = TRUE,
  email_to,
  email_subject = "fysan subject",
  email_body = "fysan send",
  email_cc = "",
  email_bcc = "",
  max_bytes = 5e+07
)

fysan(
  batch = 3,
  interval = 1,
  fyle_location = ".",
  fyle_extension = NULL,
  fyle_exclusions = NULL,
  list_recursive = TRUE,
  list_full_names = TRUE,
  email_to,
  email_subject = "fysan subject",
  email_body = "fysan send",
  email_cc = "",
  email_bcc = "",
  max_bytes = 5e+07
)
}
\arguments{
\item{batch}{How many attachments per email. Defaults to 3.}

\item{interval}{How many seconds to wait between each email send. Defaults to 1.  Can be handy to keep from overloading.}

\item{fyle_location}{Location of directory to search. Will default to search recursively. Required.}

\item{fyle_extension}{The extension(s) you want to search for, e.g. \emph{c("txt", "R")} will send files that end in .txt or .R}

\item{fyle_exclusions}{Files with this string anywhere in the file name or location will not be sent.  Optional.}

\item{list_recursive}{Set to FALSE if you do not want to search through children of \emph{fyle_location}. Generally keep TRUE.}

\item{list_full_names}{List the full names of the location. Generally keep TRUE}

\item{email_to}{Email address to send the files to}

\item{email_subject}{Optional email subject}

\item{email_body}{Optional email body}

\item{email_cc}{Optional email cc}

\item{email_bcc}{Optional email bcc}

\item{max_bytes}{Maximum byte size. Emails with attachments exceeding this size will have the attachments split up. Those files will need to be glued together by TBD function.}
}
\description{
Uses RDCOMClient package.  Will send email from default Outlook email address.

Uses RDCOMClient package.  Will send email from default Outlook email address.
}
\examples{
fysan(batch = 3,
         email_to = "noone@nowhere.com",
         fyle_location = getwd(),
         fyle_extension = c("R", "md"))
fysan(batch = 3,
         email_to = "noone@nowhere.com",
         fyle_location = getwd(),
         fyle_extension = c("R", "md"))
}
\keyword{attachments,}
\keyword{email,}
\keyword{outlook}
